// frontend/lib/api/generated/model/inspeccionDetailOut.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SIGA-UNMSM API
 * API del Sistema Integral de Gestión de Almacén y Abastecimiento Alimentario — UNMSM. Módulos activos: Autenticación, Usuarios, Almacenes, Catálogo Nutricional.
 * OpenAPI spec version: 0.1.0
 */
import type { InspeccionDetalleOut } from './inspeccionDetalleOut';

export interface InspeccionDetailOut {
  inspeccion_id: number;
  guia_remision_id: number;
  inspector_id: number;
  fecha_inspeccion: string;
  detalle: InspeccionDetalleOut[];
}