// frontend/lib/api/generated/model/ubicacionInternaOut.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SIGA-UNMSM API
 * API del Sistema Integral de Gestión de Almacén y Abastecimiento Alimentario — UNMSM. Módulos activos: Autenticación, Usuarios, Almacenes, Catálogo Nutricional.
 * OpenAPI spec version: 0.1.0
 */

export interface UbicacionInternaOut {
  ubicacion_id: number;
  almacen_id: number;
  zona: string;
  estante: string | null;
  nivel: string | null;
  contenedor_camara: string | null;
  es_cadena_frio: boolean;
}